+++ main/api.js
@@ -1,11 +1,13 @@
 const axios = require("axios");
 
-//In case update not works directly fetch from url
-async function api(url) {
-  const res = await axios.get(
-    url === "" ? "https://www.reddit.com/r/memes/new.json" : url
-  );
-  const memes = res.data.data;
-  return memes;
+async function api(subreddit, count) {
+  const url =
+    `https://meme-api.com/gimme/${subreddit}/${count}`;
+
+  console.log(url);
+
+  const res = await axios.get(url);
+
+  return res.data;
 }
 module.exports = { api };