+++ test/suite/index.js
@@ -1,9 +1,8 @@
 const path = require("path");
 const Mocha = require("mocha");
-const glob = require("glob");
+const { glob } = require("glob");
 
 function run() {
-  // Create the mocha test
   const mocha = new Mocha({
     ui: "tdd",
     color: true,
@@ -11,30 +10,22 @@ function run() {
 
   const testsRoot = path.resolve(__dirname, "..");
 
-  return new Promise((c, e) => {
-    glob("**/**.test.js", { cwd: testsRoot }, (err, files) => {
-      if (err) {
-        return e(err);
-      }
+  return glob("**/*.test.js", { cwd: testsRoot })
+    .then((files) => {
+      files.forEach((f) => {
+        mocha.addFile(path.resolve(testsRoot, f));
+      });
 
-      // Add files to the test suite
-      files.forEach((f) => mocha.addFile(path.resolve(testsRoot, f)));
-
-      try {
-        // Run the mocha test
+      return new Promise((resolve, reject) => {
         mocha.run((failures) => {
           if (failures > 0) {
-            e(new Error(`${failures} tests failed.`));
+            reject(new Error(`${failures} tests failed.`));
           } else {
-            c();
+            resolve();
           }
         });
-      } catch (err) {
-        console.error(err);
-        e(err);
-      }
+      });
     });
-  });
 }
 
 module.exports = { run };